pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "ARIA — Autonomous Reasoning & Intelligent Assistant"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Fulvio", email = "fulvio@example.com" }
]
keywords = ["ai", "agent", "assistant", "autonomous", "mcp"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core models and validation
    "pydantic>=2.6,<3.0",

    # MCP server
    "fastmcp>=3.2,<4.0",

    # Telegram bot (async)
    "python-telegram-bot>=22.0,<23.0",

    # Database
    "aiosqlite>=0.19",

    # Cryptography and secrets (sops CLI is used directly, age for encryption)
    "cryptography>=42.0",
    "keyring>=25.0",
    "secretstorage>=3.3",

    # HTTP client
    "httpx>=0.27",

    # Resilience patterns
    "tenacity>=8.2",  # retry/circuit breaker

    # Structured logging
    "structlog>=24.1,<26.0",

    # YAML config
    "PyYAML>=6.0,<7.0",

    # Scheduling
    "croniter>=2.0",
    "python-dateutil>=2.8",

    # CLI and output
    "rich>=13.0",
    "typer>=0.12",

    # Systemd integration
    "sd-notify>=0.1",
]

[project.optional-dependencies]
dev = [
    # Linting and formatting
    "ruff>=0.9",

    # Type checking
    "mypy>=1.15",

    # Testing
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",

    # Proxy dev deps (MCP tool search)
    "respx>=0.21",
    "numpy>=1.26",
]

ml = [
    "lancedb>=0.30,<0.31",
    "faster-whisper>=1.2,<2.0",
    "pillow>=10.0",
    "pytesseract>=0.3",
]

[project.scripts]
aria = "aria.bin.aria:main"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aria"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/CONTRIBUTING.md",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "ASYNC",  # flake8-async
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TCH", # flake8-typecheck
    "ARG", # flake8-unused-arguments
    "PLE", # pydantic-lint
    "PLR", # pydantic-lint
    "PLW", # pydantic-lint
]
ignore = [
    "PLR0913", # too-many-arguments
    "PLR2004", # magic-value-comparison
    "RET505",  # unnecessary-else-after-return
    "ISC001",  # string-concatenation (conflicts with formatter)
    "ARG002",  # unused-method-argument (stub code)
    "ARG003",  # unused-class-argument
    "ARG005",  # unused-function-argument
    "PIE790",  # unnecessary-pass-statement (stub code)
    "ARG001",  # unused-function-argument (stub code)
]

[tool.ruff.lint.per-file-ignores]
"src/aria/memory/migrations.py" = ["E501"]
"src/aria/memory/semantic.py" = ["E501"]
"src/aria/memory/wiki/db.py" = ["TC003"]
"scripts/*.py" = ["ANN201", "ANN002", "ANN401", "ASYNC240", "B007", "DTZ003", "E402", "E501", "F841", "N999", "PIE810", "PLR0912", "PLR0915", "SIM102", "SIM105", "SIM108", "T201"]
"scripts/benchmarks/*.py" = ["ANN201", "ANN002", "ANN401", "ASYNC240", "B007", "DTZ003", "E402", "E501", "F841", "N999", "PIE810", "PLR0912", "PLR0915", "SIM102", "SIM105", "SIM108", "T201"]
"scripts/check_mcp_drift.py" = ["T201"]
"scripts/workspace-write-health.py" = ["T201"]
"scripts/workspace_auth.py" = ["T201"]
"tests/benchmarks/*.py" = ["T201"]
"tests/**/*.py" = ["ANN", "TC003"]
"tests/integration/agents/search/*.py" = ["E501"]
"tests/integration/mcp/proxy/*.py" = ["E501"]
"tests/unit/agents/search/*.py" = ["E501", "F841"]
"tests/unit/credentials/*.py" = ["PLW0108", "TC002"]
"tests/unit/mcp/proxy/test_middleware.py" = ["E702"]
"tests/unit/memory/test_mcp_server.py" = ["PLW0108"]
"tests/unit/memory/wiki/*.py" = ["B017", "E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "pydantic.*",
    "pydantic_settings.*",
    "structlog",
    "prometheus_client",
    "croniter",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src/aria"]
branch = true
omit = [
    "**/tests/**",
    "**/__pycache__/**",
    "**/.venv/**",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "numpy>=1.26",
    "respx>=0.21",
]
